Worked on micro settings for both explorers.
Implemented a configuration function, which acts to open up a yaml file
containing settings. this will then be used by multiple functions. I
have commented out the global variables which were used within the
functions, and replaced with the new method. The rest were only used in
__name__ == main section. We can implement this later if necessary. I
also updated some of twitter_explorer (to make use of the setting which
allows user to store all sources)